catalog: Extracted CatalogMongoSampleDBAdaptor from CatalogMongoDBAdaptor
CatalogMongoSampleDBAdaptor extends CatalogDBAdaptor implements CatalogSampleDBAdaptor
